Global: Make notes and examples conform, maybe
One thing left over:  Any instance of a NOTE 1 that does not have
a NOTE 2 must omit the number, so "NOTE" rather than "NOTE 1".
Unless someone has some bright ideas, this is a manual operation.
Maybe like this for maintainability:

	%\begin{note}
	[NOTE: .... ]%\end{note}

I don't know if "NOTE" is supposed to be enclosed in "[]".  I am
assuming that it is.

I don't know if the "--- end EXAMPLE]" is OK.  I can make it work
like the note environment if that is the intent.

Signed-off-by: Paul E. McKenney <paulmck@kernel.org>